+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_job/job_service.py
@@ -17,7 +17,6 @@ from azure.ai.ml.exceptions import ErrorCategory, ErrorTarget, ValidationErrorTy
 module_logger = logging.getLogger(__name__)
 
 
-@experimental
 class JobServiceBase(RestTranslatableMixin, DictMixin):
     """Base class for job service configuration.
 
@@ -148,7 +147,6 @@ class JobServiceBase(RestTranslatableMixin, DictMixin):
         return result
 
 
-@experimental
 class JobService(JobServiceBase):
     """Basic job service configuration for backward compatibility. Use JupyterLabJobService, SshJobService,
     TensorBoardJobService or VsCodeJobService specific to your job type instead of JobService.
@@ -178,7 +176,6 @@ class JobService(JobServiceBase):
         return self._to_rest_job_service()
 
 
-@experimental
 class SshJobService(JobServiceBase):
     """SSH job service configuration.
 
@@ -233,7 +230,6 @@ class SshJobService(JobServiceBase):
         return self._to_rest_job_service(updated_properties)
 
 
-@experimental
 class TensorBoardJobService(JobServiceBase):
     """TensorBoard job service configuration.
 
@@ -288,7 +284,6 @@ class TensorBoardJobService(JobServiceBase):
         return self._to_rest_job_service(updated_properties)
 
 
-@experimental
 class JupyterLabJobService(JobServiceBase):
     """JupyterLab job service configuration.
 
@@ -336,7 +331,6 @@ class JupyterLabJobService(JobServiceBase):
         return self._to_rest_job_service()
 
 
-@experimental
 class VsCodeJobService(JobServiceBase):
     """VS Code job service configuration.